+++ assembly/hittables/Hittable.ts
@@ -5,8 +5,8 @@ export default abstract class Hittable
 {
   public abstract hit (
     ray: Ray,
-    tMin: f64,
-    tMax: f64,
+    tMin: f32,
+    tMax: f32,
     record: Record
   ): bool;
 }
